Verilog Parsing: Reading in Meta Data
Now we're properly parsing in all the attributes and
the components are already kinda placed.

There is still some debugging to do and the wires
aren't properly placed yet, but we're getting somewhere.

Close the file after we parsed

Add .vs files to the list of files which will be treated as
a schematic instead of a text file