Verilog Parsing: Solving the positioning issue
By loading the orientation info first and borrowing some
code from the recreate method of the Component class,
the orientations look good now, so far.

However. The net names and component names with the Verilog
requirement of having asterisks and dots prefixed with a \
still can not be loaded properly.